UPDATE : Refactor RentCar
