+++ src/main/java/com/woowa/rentcar/view/InputView.java
@@ -8,18 +8,10 @@ import java.util.Scanner;
 public class InputView {
     private static Scanner scanner = new Scanner(System.in);
 
-    public static void title() {
-        System.out.println("어서오세요. 우아한 렌트카 입니다.");
-        System.out.println("렌트 가능 차종은 다음과 같습니다.");
-        System.out.println("1. Sonata");
-        System.out.println("2. Avante");
-        System.out.println("3. K5");
-    }
-
     public static CarName car() {
-        System.out.print("렌트할 차종을 입력해주세요.\n(추가할 차가 없으면 end 를 입력) : ");
+        OutputView.car();
         try {
-            return new CarName(scanner.next());
+            return new CarName(scanner.nextLine());
         } catch (IllegalArgumentException e) {
             System.out.println(e.getMessage());
             return car();
@@ -27,9 +19,9 @@ public class InputView {
     }
 
     public static Distance distance() {
-        System.out.print("예상 주행 거리를 입력해주세요 : ");
+        OutputView.distance();
         try {
-            return new Distance(scanner.next());
+            return new Distance(scanner.nextLine());
         } catch (IllegalArgumentException e) {
             System.out.println(e.getMessage());
             return distance();
